more blackjack logic (probably moving to a different project eventually)
