+++ app/States/Blackjack/DealerState.php
@@ -9,4 +9,6 @@ class DealerState extends State
     public array $hand;
 
     public string $deck;
+
+    public int $game_id;
 }
